api 예제 삭제

+++ src/main/java/hello/core/member/MemberService.java
@@ -4,5 +4,5 @@ public interface MemberService {
 
     void join(Member member);
 
-    Member findMember(Long memberId);
+    Member findMember(int memberId);
 }
